Demo malaria datasets

These datasets are simulated data representing monthly malaria cases and associated covariates in the District of Ifanadiana, Vatovavy, Madagascar. They represent 21 orgUnits and 106 months of data (October 2016 - July 2024). They are used in demos and in internal testing.

Usage

demo_malaria

Format

demo_malaria

A data frame with 2226 rows and 12 columns:

orgUnit

DHIS2 orgUnit corresponding to primary health clinics

n_case

Number of monthly malaria cases

pev

Average monthly potential evapotranspiration

rain_mm

Total monthly rainfall

temp_c

Average monthly temperature

wealth_index

Socio-economic index

elevation

Elevation

LLIN_use

Proportion of households with at least two LLINs per member

csb_type

Type of clinic: CSB1 or CSB2

time_to_district

Travel time to district capital from clinic

LLIN_wane

Number of years since LLIN distribution campaign

period

Year and month followin DHIS2 period style (YYYYMM)