Estimate variable importance and partial dependence plots of a ranger model

Estimate variable importance and partial dependence plots of a ranger model

Usage

inv_variables_ranger(
  cv_set,
  y_var,
  id_vars,
  pred_vars,
  hyper_control = list(mtry = NULL, min.node.size = NULL, num.trees = 500),
  var_scales
)

Arguments

cv_set

list object containing analysis and assessment data.frames

y_var

string name of column of observed values

id_vars

names of columns used as keys/ids in data.frames

pred_vars

vector string of predictor variables to use in model

hyper_control

hyperparameters to use in the model, in either list or named vector format. Defaults are ranger defaults

var_scales

data.frame containing centering and scaling parameters for variables

Value

list containing variable importance scores and a list of dataframes containing data for pdp plots with each element corresponding to a variable