Run full train_model workflow using multiple modeling approaches

Run full train_model workflow using multiple modeling approaches

Usage

train_models(
  prep_output,
  models,
  y_var,
  pred_vars,
  id_vars = c("orgUnit", "date"),
  results_dir = NULL,
  tune = NULL,
  model_configs
)

Arguments

prep_output

output of prep_data

models

vector of models to fit. Options: naive, arimax, glm_nb, ranger, inla

y_var

character. name of variable to predict

pred_vars

character vector of predictor variables. Not all will be used in all models

results_dir

directory to save model results and quarto document in. If NULL, temporary directory will be used

tune

string vector containing names of models to tune. Options = ranger

model_configs

Optional list of configurations for each model. Name of element in list should match model name. See ensemble_forecast for more info.

Value

saves model outputs to results_dir if specified. Creates a quarto doc of model outputs